requirements.txt
# Core
torch>=2.0.0
torchaudio>=2.0.0

# TTS engines
TTS>=0.22.0
kokoro>=0.9.4
transformers>=4.40.0
soundfile>=0.12.1
scipy>=1.11.0

# STT
openai-whisper>=20231117

# Language detection
fasttext-wheel>=0.9.2

# LLM
sarvamai>=0.1.0

# NLP / Translation
sentencepiece>=0.1.99
sacremoses>=0.0.53

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Audio recording (assistant UI)
pyaudio>=0.2.14
numpy>=1.24.0
